chore(api): cleanup PR6 surface before staging — renames, deprecations, filter tightening
Builds on the two prior PR6-hardening commits. No behavior change to the
canonical resolvers; this commit closes review gaps that block merging
to staging:

- Rename pr5/pr6-scaffolded files and describe blocks (one-shot cleanup
  of staged-rollout names). Split pr6.ts into activity.ts + leaderboard.ts
  so the two resolvers live in their own files alongside their tests.
- @deprecate accountAccuracyRank, accountStats, accountStatsLeaderboardPage,
  accountStatsRank, accuracyLeaderboardPage, activityPage so the migration
  window opens with the new Relay-shaped surface as the canonical path.
- Promote Account.address from String! to Address! (additive scalar
  tightening — wire format unchanged).
- Reject unsupported timestamp operators in Account.stats and
  leaderboard filters with clear errors rather than silently dropping them.
- ActivityFilter: add createdAt: IntFilter, conditionIds: [String!], and
  conditionGroupId: Int so external clients can use what the resolver
  already accepts. Switch LeaderboardFilter from {from, to} to
  operator-pattern {timestamp: IntFilter}.
- PredictionFilter / ForecastFilter: expose conditionIds: [String!] so
  the SDL matches what predictionsConnection / forecastsConnection
  already accept (closing the group-backed Question silent-no-op gap).
- Drop the AccountFilter SDL input (unreferenced — there is no
  accountsConnection yet).
- Fix three stale test assertions:
  - globalId.test.ts FROZEN_NODE_TYPES list now includes Account
  - Condition.test.ts drops Condition.predictions DataLoader tests (the
    field now returns PredictionConnection!; behavior covered in
    crossStream.test.ts)
  - rename pr5/pr6 describe blocks to plain names
